refactor: merge all components into a single file.
